Implement subtests properly and clean up various issues
Added a comment at the end of all suites if any tests failed.

When a subtest ends send a Test event to the next suite up the stack to
indicate whether or not the subtest passed.

Added subtest tests.

Removed the timing output for now. For now I'll aim at replicating what Perl 5
outputs.

Moved pluralization to its own sub.