Implement a Test::Stream::Suite object
This is the beginning of a refactor to add a suite stack to the hub (I may
move it again later). Previously I was doing this in the TAP12 formatter class
but I realized that tracking the stack of suites is required regardless of the
formatter, since a sub-suite is counted as a single test in the containing
suite, and so the containing suite needs to know it's pass/fail status.